+++ src/highlighting.js
@@ -115,26 +115,4 @@ function highlightBml(source) {
   return hljs.highlightAuto(source, ['bml', 'bml_body_only']).value;
 }
 
-let bmlSrc = '' + fs.readFileSync('./highlight_sample_2.bml');
-let css = '' + fs.readFileSync('./node_modules/highlight.js/styles/darcula.css');
-
-
-let highlighted = highlightBml(bmlSrc);
-
-let output = `
-<html>
-<style>
-${css}
-</style>
-
-<body>
-<pre>
-<code id="code-block" class="hljs">
-${highlighted}
-</code>
-</pre>
-</body>
-</html>
-`;
-
-fs.writeFileSync('out.html', output);
+exports.highlightBml = highlightBml;